update tests for trait parameter attribute removal
